component interaction with image loader completed

+++ src/app/image-loader/image-loader.component.ts
@@ -1,7 +1,9 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 import { FileMetadata } from './file-metadata';
 import { MatImageLoaderService } from './image-loader.service'; 
+import { Observable } from 'rxjs/Observable';
+//import { EventEmitter } from 'protractor';
 
 @Component({
   selector: 'mat-image-loader',
@@ -20,60 +22,73 @@ export class MatImageLoaderComponent implements OnInit {
   ngOnInit() {
   }
 
+  public getFiles() : FileMetadata[] {
+    return this.uploadingFiles;
+  }
+
+  public get isUploadCompleted(): boolean {
+    let pending = this.uploadingFiles.filter(function(entry:FileMetadata){
+      return entry.subscription || entry.subscription != null;
+    });
+
+    return !pending || pending.length == 0;
+  } 
+
   public fileChanged(event) {
     let fileList: FileList = event.target.files;
 
     if (fileList.length > 0) {
-      let file: File = fileList[fileList.length - 1];
+      const file: File = fileList[fileList.length - 1];
       this.pushNewFile(file);
-      this.imageLoaderService.createFile(file, this.apiEndPoint)
-        .subscribe(
-          data => {
-            const fileMetadata = this.uploadingFiles.find(function (value: FileMetadata) {
-              return value.description === file.name
-                && !value.id;
-            });
+      const fileMetadata = this.uploadingFiles.find(function (value: FileMetadata) {
+        return value.description === file.name
+          && !value.id;
+      });
 
+      let observable = this.imageLoaderService.createFile(file, this.apiEndPoint);
+
+      fileMetadata.subscription = observable.subscribe(
+          data => {
             event.target.value = '';
 
             fileMetadata.id = data.id;
             fileMetadata.dataUrl = this.apiEndPoint + '/' + data.id;
+            delete fileMetadata.subscription;
           },
           error => {
             console.log(error);
             event.target.value = '';
+            delete fileMetadata.subscription;
           }
         );
     }
   }
 
   public deleteFile(id: number) {
-    this.imageLoaderService.deleteFile(id, this.apiEndPoint).subscribe(
-      success => {
-        let oDeleted = this.uploadingFiles.find(function(value: FileMetadata, index: number, obj: FileMetadata[] ) {
-          return value.id === id;
-        });
+    let oDeleted = this.uploadingFiles.find(function(value: FileMetadata, index: number, obj: FileMetadata[] ) {
+      return value.id === id;
+    });
 
-        this.uploadingFiles.splice(this.uploadingFiles.indexOf(oDeleted), 1);
-      },
-      error => console.log(error));
+    if (oDeleted.subscription) {
+      //Upload of file is not finished
+      oDeleted.subscription.unsubscribe();
+      this.uploadingFiles.splice(this.uploadingFiles.indexOf(oDeleted), 1);
+    } else {
+      this.imageLoaderService.deleteFile(id, this.apiEndPoint).subscribe(
+        success => {
+  
+          this.uploadingFiles.splice(this.uploadingFiles.indexOf(oDeleted), 1);
+        },
+        error => console.log(error));
+    }
   }
 
   private pushNewFile(file: File) {
     let fileMeta: FileMetadata = {};
-    //let myReader: FileReader = new FileReader();
-    //let that: MatImageLoaderComponent = this;
 
     fileMeta.description = file.name;
 
-    //myReader.onloadend = function (e) {
-      // you can perform an action with readed data here
-     // fileMeta.dataUrl = myReader.result
       this.uploadingFiles.push(fileMeta);
-    //}
-
-    //myReader.readAsDataURL(file);
-
   }
 
 }
